Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/websockets/Send-unicode-data.any.worker-expected.txt


FAIL Send unicode data on a WebSocket - Connection should be opened assert_equals: expected 0 but got 12
PASS Send unicode data on a WebSocket - Message should be received
FAIL Send unicode data on a WebSocket - Connection should be closed assert_true: WebSocket connection should be open expected true got false
